move DoSFilter to jetty pkg
